Add config read-api and use it to detect public access
Currently no way to modify the settings except by manual SQL, but it's
coming.

This allows us to have event-specific configuration in the database. This
commit adds js that uses this to detect if the frontend is a public vhost
or not.

There are currently only three columns provided in addition to ID.

publicvhost: The domain name of the publicly accessible version of Gondul.
shortname: code name for the vent. e.g dx16, tg17, etc.
data: jsonb to contain most other settings as we see fit.

Most settings will be stored in 'data' if it's only or mainly used by the
frontend. This will allow frontend developers to supply arbitrary
configuration options without the need for API or database changes in the
future.

Fixes #51

References #54

Still need the GUI for it.