front: Move data-logic out of infobox
handlerInfo is getting slightly extensive, but it might be worth it.

This also introduces a "map" handler that .... doesn't handle the map.

If/when namespaces are introduced properly in nms-map-handlers.js, we
should also rename it to nms-data-logic.js or something else that signifies
that this is what makes data useful one way or an other.