Fix numerous time-travel issues in front and API
Fixes #69 #11 #5

Introduces nmsTime which unifies the time travel code a bit. It still needs
some work, but this is much better.

All conversion is now done by native JavaScript methods, freeing us from
the hell that is parsing it ourself.

One thing should be added: The backend should discard any now=values that
are not 5-minute intervals. We don't want to kill the cache and the
database by extension.

Still need to re-implement the "replay event" shorthand, but that ties in
to #54