front: Greatly improve health map and more
As a result of testing with TG16-data, I was able to find and fix a few
outstanding issues and tune the health map to make more sense.

Also introduced both temperature and dhcp into the health map and info box.

Removed the switchInfo text on the health map unless the score is above a
threshold.

Still more tuning needed. After introducing dhcpInfo it's become even more
important to address #88