api: Fix bad hash-logic, front: Performance and more
Fixes #93
Fixes #92
Fixes #91

The hash logic with FreezeThaw wasn't consistent, which broke a lot of
optimizations. This might be slower (?) but it's consistent which means
we'll make up for it by better cache utilization.

To utilize this I also added delayed JSON parsing (#91) with a big
description of why we want it. This also means that #90 and #89 is a lot
less important. One issue right now is that "time" isn't part of the ETag.
This is both a blessing and a curse. If we are going through time we will
actually use in-app cache (not browser cache - the URL changes). This
makes for nice CPU savings, but ultimately means we can't tell that the
data for two distinct times is identical because we only see one. Oh
well...

On the simpler side, I fixed/improved the clearBox logic to fix the
artifacts. Might need to tune this. I also added some optimizations to
avoid redrawing identical text. This is what pointed out to me that the
hash was broken too: The resizeEvent was triggering all the time, forcing
redraws all the time. This happened because nmsData.switches "updated"
frequently which should only happen if someone adds/moves/removes a switch
or linknet - regardless of time travel. This was caused by the inconsistent
hash causing a false cache miss.

Anyway, if you are still reading this, you almost got a patch with "kjell
magne bondevik uten mellomnavn" as a literal text string... TOO BAD.