tools: redo discovery in the form of lolwhat
what?

lol...

just figure it out for me.

Ok, so the tool isn't entirely done, but it departs from the simple
lldpdiscver.pl method that was very limited when things weren't working as
they should.

The long-term idea of lolwhat is that it should be able to do all sorts of
discovery. Right now it just takes a list of IPs and figures out how things
are connected by SNMP. It uses LLDP info from SNMP, but it does not
_currently_ traverse new hosts based on what it finds.