lolwhat + snmp: Performance tweaks and better collision detection
Truns out people suck at making sensible network hardware.

Lots of replies on IPs that the devices didn't know they had. I discard
these now, but should keep some sort of map. The reason this is a problem
is that we later "Deduplicate" results by comparing IP lists - but that
doesn't work when the devices don't know about all of the IPs they listen
to.

I might have to revisit this since the IPs are obviously valid...