lolwhat: Add deep comparison of duplicate results
When a device replies from two or more IPs we can't assume it actually
knows what IPs it replied from... (great, right?).

This adds a fall-back when that happens where we compare the actual
interfaces. We require a minimum of 10 matching MAC addresses and skip the
once that are blank to gain confidence in the result.