DOCKER + ANSIBLE == PAIN (also: graphite)
3 hours.

THREE HOURS.

The extra "Expose" statements are to workaround an issue introduced in
Ansible 2.1 with the addition of the docker_* modules that replaces the
"docker" module.

THREEEEEEE HOURS.

The graphite container is so far not ued for anything, but is provided to
get things rolling. It does do persistent storage, but obviously not the
way I originally wanted. Because persistent storage with docker is a
pile of frozen piss.