Use graphite bandwidth graphs in snmp and nuke "traffic"
The "traffic" thing should be integrated in the snmp/ports overview.

All that's missing is a two-step drill-down per port. E.g.: Click a port
and you get graphs, click the next drill down and get full snmp-data.

When that's done, all we need is an "expand all" button.

And a pony.