Introduce nms-ui-boxes, a big step for GUI
nmsBox is a new class for generic HTML-based elements, and will eventually be
used to replace nms-info-box.

The idea is simple: a generic way to deal with containers that are usually
represented in HTML too.

To test it, I've re-implemented the GUI for the oplog. This rewrite didn't
really utilize the new benefits of the framework, but was a small step.

I've also added nms-ui-switch, which isn't exposed anywhere yet and isn't
done, but is a good start.

nms-ui-switch will be the new way to add and edit switches in the future, it
will probably be made more generic over time and thus can be reused for
networks too.

Note how x = nmsUiSwitch() will allow you to do x.row["community"].value
to both get and set the value, and setting will visually update and run
any verifier that will be relevant, and alert the parent. This can then
be used for simple stuff like json-verification, but also for stuff like
auto-complete or whatnot. God only knows.

Obviously I will continue to work on this over the next few days...