Extenstive frontend work on box/type/switch editing
Still not very visible, but the nmsModSwitch() class is starting to look
very interesting.

This introduces a big thing: A type system. So far it is "only" used to
provide generic validation-services for all fields that will be edited, but
the scope can extend further into other uses.

The implementation as it is does shallow validation for IP addresses,
JSON-input (tags,placement), durations/intervals (poll frequency), sysname
references (checks that distro_name is a valid system), and networks
(verifying that mgmt_vlan / traffic_vlan exists). This is expected to
improve over time, and should eventually lead to things like "if it doesn't
exist, then please make it."

As the big comment in nms-ui-switch.js now suggests: This is still just a
proof of concept and while the fundamental concepts of nmsBox seems to work
well, I still feel like I have a lot to iron out in nmsModSwitch, which
really should be generic for networks and switches at the very least.

But we're getting there.