Expose new-style edit-stuff in old-style panel
The integration is a hack, but a first step. I will eventually remove all the
code in nms-info-box, but I needed a simple test.

This also demonstrates how to add a new switch.... simply hit edit, then change
the name. It aint pretty, and shouldn't work like it does today, but it's
a decent example.

Also, had to fix the backend again now that I actually tested the write-api :D