front: Implement new network/switch and more
Fixes #183
References #182

It's now possible to add new networks and switches, and it's also possible to
edit networks, but since I haven't exposed it in the UI yet, I can't really
close the bug.

This is also a major step in stripping nms-info-box apart.

Next up is listing existing networks, possibly linking to them. I think that
might be a decent first-step towards a new core nms-info-box since it doesn't
alreay exist.