Add basic collector-implementation
References #195

This is the first dumb-ass implementation of the collector-api.

It actually works fine for write AS IS, but has 0 safties. But I'm thinking the
API itself is reasonably stable and ready for testing.

I will need to extend it for dhcp-specific stuff, and the plan is to
create a secondary table with metadata (hidden from the user) and a few
other nice things.

Also, indexes. We need indexes. Obviously.