fix(nightmode): Respect user provided configuration for night mode if provided
If user sets the cookie or URL query parameter for nightMode, we will
respect that. Note that it is a bit cumbersome to _clear_ this
configuration (afaik?) if it is set.

This is also the case if the user toggles nightmode directly in Gondul
(`n`).
However, this last configuration is not persisted - so on a page reload,
this setting is removed (no need to clear cookies/query params), and
the system preference is again followed.